+++ R/RcppExports.R
@@ -37,7 +37,7 @@
 #' 
 #' @references
 #' Tillé, Y., Dickson, M.M., Espa, G., and Guiliani, D. (2018). Measuring the spatial balance of a sample: A new measure based on Moran's I index.
-#' \emph{Spatial Statistics}, 23, 182-192. \url{https://doi.org/10.1016/j.spasta.2018.02.001}
+#' \emph{Spatial Statistics}, 23, 182-192.
 #' 
 #' @seealso
 #' \code{\link{wpik}}
@@ -436,7 +436,7 @@ wpik <- function(X, pik, bound = 1.0, tore = FALSE, shift = FALSE, toreBound = -
 #' 
 #' @references 
 #' Tillé, Y., Dickson, M.M., Espa, G., and Guiliani, D. (2018). Measuring the spatial balance of a sample: A new measure based on Moran's I index.
-#' \emph{Spatial Statistics}, 23, 182-192. \url{https://doi.org/10.1016/j.spasta.2018.02.001}
+#' \emph{Spatial Statistics}, 23, 182-192.
 #' 
 #' @seealso
 #' \code{\link{wpik}}, \code{\link{distUnitk}}, \code{\link{wave}}.